requirements.txt
aioredis==1.3.1
amqp==5.0.2
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
billiard==3.6.3.0
celery==5.0.2
cffi==1.14.4
channels==2.4.0
channels-redis==3.2.0
chasgimqtt @ git+https://github.com/ulises58/channels-asgi-mqtt.git@c47352bead3495db569cd84db5de2d601955856a
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
constantly==15.1.0
cryptography==3.2.1
daphne==2.5.0
Django==3.1.4
django-environ==0.4.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
kombu==5.0.2
msgpack==1.0.0
mysqlclient==2.0.1
paho-mqtt==1.5.1
prompt-toolkit==3.0.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.0
pytz==2020.4
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.4.1
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.2.0
